Complete rewrite of the UI code to make it KDE standards compliant
svn path=/trunk/kdegraphics/kcoloredit/; revision=159688

+++ kzcolorselector.cpp
@@ -119,7 +119,7 @@ void KZColorSelector::setColor(QColor* const color, const int y) {
 }
 
 QColor* KZColorSelector::getStandardColorsPalette() {
-	QColor* palette = new QColor[STANDARD_PALETTE_SIZE];
+	QColor* palette = new QColor[( int )STANDARD_PALETTE_SIZE];
 	int i = 0;
 	palette[i++] = Qt::red;
 	palette[i++] = Qt::green;
